services implemented, UI template added

+++ Services/MoneySaver.Services.Data/CategoriesService.cs
@@ -20,11 +20,24 @@
             this.dbContext = dbContext;
         }
 
-        public async Task<string> AddAsync(string categoryName)
+        public async Task<string> AddAsync(string categoryName, int walletId)
         {
+            Wallet wallet = await this.dbContext.Wallets.FirstOrDefaultAsync(x => x.Id == walletId);
+
+            if (wallet == null)
+            {
+                throw new ArgumentException(GlobalConstants.WalletNotExist);
+            }
+
+            if (await this.IsCategoryExistAsync(walletId, categoryName))
+            {
+                throw new ArgumentException(GlobalConstants.ExistingCategory);
+            }
+
             Category category = new Category
             {
                 Name = categoryName,
+                WalletId = wallet.Id,
             };
 
             await this.dbContext.Categories.AddAsync(category);
@@ -33,15 +46,21 @@
             return string.Format(GlobalConstants.SuccessfullyAddedCategory, category.Name);
         }
 
-        public async Task<CategoryInfoDto> GetCategoryAsync(int categoryId)
+        public async Task<CategoryWalletInfoDto> GetCategoryAsync(int categoryId, int walletId)
         {
-            CategoryInfoDto category = await this.dbContext.Categories
-                .Select(categ => new CategoryInfoDto
+            CategoryWalletInfoDto category = await this.dbContext.WalletsCategories
+                .Where(x => x.CategoryId == categoryId && x.WalletId == walletId)
+                .Select(categ => new CategoryWalletInfoDto
                 {
-                    Name = categ.Name,
-                    Id = categ.Id,
+                    CategoryName = categ.Category.Name,
+                    WalletName = categ.Wallet.Name,
                 })
-                 .FirstOrDefaultAsync(x => x.Id == categoryId);
+                .FirstOrDefaultAsync();
+
+            if (category == null)
+            {
+                throw new ArgumentException(GlobalConstants.UnexistingCategory);
+            }
 
             return category;
         }
@@ -49,9 +68,22 @@
         public async Task<string> RemoveAsync(int categoryId)
         {
             Category category = await this.dbContext.Categories.FirstOrDefaultAsync(x => x.Id == categoryId);
+
+            if (category == null)
+            {
+                throw new ArgumentException(GlobalConstants.UnexistingCategory);
+            }
+
             this.dbContext.Categories.Remove(category);
+            await this.dbContext.SaveChangesAsync();
 
             return string.Format(GlobalConstants.SuccessfullyRemovedCategory, category.Name);
+        }
+
+        private async Task<bool> IsCategoryExistAsync(int walletId, string categoryName)
+        {
+            return await this.dbContext.WalletsCategories
+                .AnyAsync(cat => cat.WalletId == walletId && cat.Category.Name == categoryName);
         }
     }
 }
